+++ automations/decodex/scripts/config/automation_sync/__init__.py
@@ -1,0 +1,1 @@
+"""Automation sync helpers for checked-in Decodex automation manifests."""
